Renaming classes, methods Minor fixes
