+++ example-tests/src/main/java/com/epam/restassured/service/client/SignUpServiceClient.java
@@ -8,19 +8,16 @@ import com.epam.restassured.url.UrlBuilder;
 import com.google.common.collect.ImmutableMap;
 import com.jayway.restassured.RestAssured;
 import com.jayway.restassured.http.ContentType;
-import com.jayway.restassured.internal.http.URIBuilder;
 import com.jayway.restassured.response.Response;
 
-import java.util.Collections;
 import java.util.Map;
-import java.util.function.Function;
 
 public class SignUpServiceClient {
     private final UrlBuilder urlBuilder = new UrlBuilder(new EnvironmentProvider().get().get("BASE_URL"));
     private final PathProvider pathProvider = new SignUpPathProvider();
 
     public Response signUp(SignUpModel model) {
-        return RestAssured.given().contentType(ContentType.JSON).post(
+        return RestAssured.given().contentType(ContentType.JSON).when().post(
                 urlBuilder.withQuery(getQueryFor(model)).buildUriFor(pathProvider));
     }
